tests/wpt/web-platform-tests/bluetooth/requestDevice/acceptAllDevices/device-with-name.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
bluetooth_test(() => {
  let device_name = 'LE Device';
  return setUpPreconnectedDevice({name: device_name})
    .then(() => requestDeviceWithTrustedClick({acceptAllDevices: true}))
    .then(device => {
      assert_equals(device.name, device_name);
    });
}, 'A device with name and no UUIDs nearby. Should be found if ' +
   'acceptAllDevices is true.');
</script>
